TINY-13110: refactor
