+++ modules/tinymce/src/core/main/ts/SelectionOverrides.ts
@@ -1,7 +1,6 @@
 import { Arr, Obj, Type, Unicode } from '@ephox/katamari';
 import { Attribute, Compare, Css, Focus, Insert, InsertAll, Remove, SelectorFilter, SelectorFind, SugarElement } from '@ephox/sugar';
 
-import { isUcVideo } from './api/dom/ControlSelection';
 import type Editor from './api/Editor';
 import VK from './api/util/VK';
 import * as CaretContainer from './caret/CaretContainer';
@@ -305,7 +304,7 @@ const SelectionOverrides = (editor: Editor): SelectionOverrides => {
   };
 
   const selectElement = (elm: HTMLElement) => {
-    const targetClone = isUcVideo(elm) ? getUcVideoClone(elm) : elm.cloneNode(true);
+    const targetClone = NodeType.isUcVideo(elm) ? getUcVideoClone(elm) : elm.cloneNode(true);
     const e = editor.dispatch('ObjectSelected', { target: elm, targetClone });
     if (e.isDefaultPrevented()) {
       return null;
