+++ modules/tinymce/src/core/main/ts/dom/NodeType.ts
@@ -113,6 +113,16 @@ const isListItem = matchNodeName<HTMLLIElement>('li');
 const isDetails = matchNodeName<HTMLDetailsElement>('details');
 const isSummary = matchNodeName<HTMLElement>('summary');
 
+const ucVideoNodeName = 'uc-video' as const;
+
+interface UcVideo extends HTMLElement {
+  nodeName: typeof ucVideoNodeName;
+  width: number;
+  height: number;
+}
+
+const isUcVideo = (el: Element): el is UcVideo => el.nodeName.toLowerCase() === ucVideoNodeName;
+
 export {
   isText,
   isElement,
@@ -144,5 +154,7 @@ export {
   isTextareaOrInput,
   isListItem,
   isDetails,
-  isSummary
+  isSummary,
+  ucVideoNodeName,
+  isUcVideo
 };
